refactor(picklist2): tweak docs, refactor utils, add simple inputs
