+++ projects/cashmere-examples/src/lib/picklist2-hc-pick-option/picklist2-hc-pick-option-example.component.html
@@ -6,7 +6,7 @@
         <hc-icon fontSet="fa" fontIcon="fa-cutlery" hcIconSm></hc-icon>&nbsp;
         {{veggie.name}}
     </hc-pick-option>
-    <hc-pick-option [value]="21">Raddish</hc-pick-option>
+    <hc-pick-option [value]="21">Radish</hc-pick-option>
 </hc-picklist2>
 <br>
