+++ projects/cashmere-examples/src/lib/picklist2-templates/picklist2-templates-example.component.html
@@ -13,6 +13,7 @@
     orphanItemsGroupName="Other Travel Methods"
     [closeGroupsByDefault]="false"
     [addCustomItem]="true">
+
      <!-- Group row template -->
      <ng-template hcPickOptgroupTmp let-item="hcOption">
         <div class="hc-pick-optgroup-inner group-row">
