+++ projects/cashmere/src/lib/util.ts
@@ -1,6 +1,7 @@
 import {Observable, Subject} from 'rxjs';
 import {takeUntil} from 'rxjs/operators';
 
+/** Parse a given input into a boolean. */
 export function parseBooleanAttribute(value: boolean | string): boolean {
     if (typeof value === 'boolean') {
         return value;
@@ -14,12 +15,34 @@ export function parseBooleanAttribute(value: boolean | string): boolean {
     throw Error(String(value) + ' is not a boolean value');
 }
 
-/* Open Source `take-until-destroy` operator from: https://github.com/NetanelBasal/ngx-take-until-destroy */
-
-function isFunction(value) {
-    return typeof value === 'function';
+/** Accepts a string, return that string with HTML tag escaped. */
+const helperEl = document.createElement('p');
+export function escapeHTML(string: string) {
+    helperEl.innerText = string; // Setting inner text causes escaping
+    return helperEl.innerHTML;
 }
 
+/** Return true if given value is not null and not undefined. */
+export function isDefined(value: any) {
+    return value !== undefined && value !== null;
+}
+
+/** Return true if given value is an object that is defined. */
+export function isObject(value: any) {
+    return typeof value === 'object' && isDefined(value);
+}
+
+/** Return true if given value is a promise. */
+export function isPromise(value: any) {
+    return value instanceof Promise;
+}
+
+/** Return true if given value is a function. */
+export function isFunction(value: any) {
+    return value instanceof Function;
+}
+
+/* Open Source `take-until-destroy` operator from: https://github.com/NetanelBasal/ngx-take-until-destroy */
 export const untilDestroyed = (componentInstance, destroyMethodName = 'ngOnDestroy') => <T>(source: Observable<T>) => {
     const originalDestroy = componentInstance[destroyMethodName];
     if (isFunction(originalDestroy) === false) {
